Disable ErrorActivity for debug builds

+++ app/src/main/java/org/kiwix/kiwixmobile/splash/SplashActivity.java
@@ -21,6 +21,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 
+import org.kiwix.kiwixmobile.BuildConfig;
 import org.kiwix.kiwixmobile.base.BaseActivity;
 import org.kiwix.kiwixmobile.error.ErrorActivity;
 import org.kiwix.kiwixmobile.intro.IntroActivity;
@@ -32,22 +33,24 @@ public class SplashActivity extends BaseActivity {
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
 
-    Context appContext = this;
-    Thread.setDefaultUncaughtExceptionHandler((paramThread, paramThrowable) -> {
+    if (!BuildConfig.DEBUG) {
+      Context appContext = this;
+      Thread.setDefaultUncaughtExceptionHandler((paramThread, paramThrowable) -> {
 
-      final Intent intent = new Intent(appContext, ErrorActivity.class);
+        final Intent intent = new Intent(appContext, ErrorActivity.class);
 
-      Bundle extras = new Bundle();
-      extras.putSerializable("exception", paramThrowable);
+        Bundle extras = new Bundle();
+        extras.putSerializable("exception", paramThrowable);
 
-      intent.putExtras(extras);
+        intent.putExtras(extras);
 
-      appContext.startActivity(intent);
+        appContext.startActivity(intent);
 
-      finish();
-      android.os.Process.killProcess(android.os.Process.myPid());
-      System.exit(10);
-    });
+        finish();
+        android.os.Process.killProcess(android.os.Process.myPid());
+        System.exit(10);
+      });
+    }
 
     Intent intent;
     if (!sharedPreferenceUtil.showIntro()) {
